app/api/routers/dependencies.py
from sqlalchemy.orm import Session
from collections.abc import Generator

from app.persistence.db import engine


def get_session() -> Generator[Session, None, None]:
    session = Session(engine)
    try:
        yield session
        session.commit()
    except Exception:
        session.rollback()
        raise
    finally:
        session.close()
