pyproject.toml
[project]
name = "clinical-lab-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.128.0",
    "freezegun>=1.5.5",
    "psycopg>=3.3.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "freezegun>=1.5.5",
    "pytest>=9.0.2",
]
